### 2015-08-21 #19 @sckott
oai - an OAI-PMH client
* 1. What does this package do? (explain in 50 words or less)

`oai` is an R client to work with OAI-PMH services, used by many libraries and other content distributors (e.g. Dryad, Datacite, etc.). The data is metadata for an object (book, video, digital files, etc.)

* 2. Paste the full DESCRIPTION file inside a code block (bounded by ``` on either end).

```
Package: oai
Title: General Purpose 'Oai-PMH' Services Client
Description: A general purpose client to work with any 'OAI-PMH'
    service. The 'OAI-PMH' protocol is described at
    http://www.openarchives.org/OAI/openarchivesprotocol.html.
    Functions are provided to work with the 'OAI-PMH' verbs: 'GetRecord',
    'Identify', 'ListIdentifiers', 'ListMetadataFormats', 'ListRecords', and
    'ListSets'.
Version: 0.0.5.9000
Authors@R: c(person("Scott", "Chamberlain", role = c("aut", "cre"),
    email = "myrmecocystus@gmail.com"))
    xml2,
    httr
```

* 3. URL for the package (the development repository, not a stylized html page)

https://github.com/sckott/oai

* 4. What data source(s) does it work with (if applicable)?

Any OAI-PMH service

* 5. Who is the target audience?

Primarily other R clients - I have a number of R clients I work on that could use this. In addition, some users may want to use this to interact with OAI-PMH services themselves.

* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?

There is an OAI-PMH client (https://cran.rstudio.com/web/packages/OAIHarvester/) on CRAN but it's built on `XML` and `RCurl`, packages basically replaced now by `xml2` and `httr`/`curl`, respectively. `oai` is built on `xml2` and `httr`. In addition, I give back tidy `dplyr` like data.frame's to make data comprehension, manipulation, and visualization easier, whereas `OAIHarvester` gives back arrays/matrices

* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
* [x] This package does not violate the Terms of Service of any service it interacts with.
* [x] The repository has continuous integration with Travis and/or another service
* [x] The package contains a vignette
* [x] The package contains a reasonably complete readme with devtools install instructions
* [x] The package contains unit tests
* [x] The package only exports functions to the NAMESPACE that are intended for end users

* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.

* [ ] Are there any package dependencies not on CRAN?
* [x] Does the package have a CRAN accepted license?
* [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.

* 9. Please add explanations below for any exceptions to the above:

* 10. If this is a resubmission following rejection, please explain the change in circumstances.
