### 2015-08-25 8866fa3
fix #2, paging works, pulled out while loop as sep helper fxn simplifying three fxns that loop bumped patch version
