+++ R/list_identifiers.R
@@ -31,38 +31,10 @@ list_identifiers <- function(url = "http://oai.datacite.org/oai", prefix = "oai_
   check_url(url)
   args <- sc(list(verb = "ListIdentifiers", metadataPrefix = prefix, from = from,
                   until = until, set = set, token = token))
-  iter <- 0
-  token <- "characters"
-  out <- list()
-  while (is.character(token)) {
-    iter <- iter + 1
-    args2 <- args
-    if (token != "characters") {
-      args2$resumptionToken <- token
-      args2$from <- NULL
-      args2$until <- NULL
-      args2$set <- NULL
-      args2$metadataPrefix <- NULL
-    }
-
-    res <- GET(url, query = args2, ...)
-    stop_for_status(res)
-    tt <- content(res, "text")
-    handle_errors(tt)
-    xml_orig <- xml2::read_xml(tt)
-    xml <- xml2::xml_children(xml2::xml_children(xml_orig)[[3]])
-    tok <- xml2::xml_text(xml2::as_list(xml[sapply(xml, xml_name) == "resumptionToken"])[[1]])
-    tok_atts <- xml2::xml_attrs(xml2::as_list(xml[sapply(xml, xml_name) == "resumptionToken"])[[1]])
-    tok <- c(token = tok, as.list(tok_atts))
-    out[[iter]] <- parse_listid(xml)
-    if (tok$token == "") {
-      token <- 1
-    } else {
-      token <- tok$token
-    }
-  }
-  out <- do.call("c", out)
-  structure(rbind_fill(out), class = c("oai_df", "data.frame"), type = "ListIdentifiers")
+  out <- while_oai(url, args, token, ...)
+  structure(rbind_fill(out),
+            class = c("oai_df", "data.frame"),
+            type = "ListIdentifiers")
 }
 
 parse_listid <- function(x) {
