### 2015-08-25 594150f
one more test added #1, list_sets egs
