+++ DESCRIPTION
@@ -7,7 +7,7 @@ Description: A general purpose client to work with any 'OAI-PMH'
     Functions are provided to work with the 'OAI-PMH' verbs: 'GetRecord',
     'Identify', 'ListIdentifiers', 'ListMetadataFormats', 'ListRecords', and
     'ListSets'.
-Version: 0.0.7.9000
+Version: 0.0.7.9400
 License: MIT + file LICENSE
 Authors@R: c(person("Scott", "Chamberlain", role = c("aut", "cre"),
     email = "myrmecocystus@gmail.com"))
