### 2015-08-25 b5e63d4
added test for id(), #1
