### 2015-08-25 d56e0eb
added load_ and update_providers, fix #6
