+++ DESCRIPTION
@@ -7,13 +7,14 @@ Description: A general purpose client to work with any 'OAI-PMH'
     Functions are provided to work with the 'OAI-PMH' verbs: 'GetRecord',
     'Identify', 'ListIdentifiers', 'ListMetadataFormats', 'ListRecords', and
     'ListSets'.
-Version: 0.0.7.9400
+Version: 0.0.7.9990
 License: MIT + file LICENSE
 Authors@R: c(person("Scott", "Chamberlain", role = c("aut", "cre"),
     email = "myrmecocystus@gmail.com"))
 URL: https://github.com/sckott/oai
 BugReports: https://github.com/sckott/oai/issues
 VignetteBuilder: knitr
+LazyData: true
 Imports:
     methods,
     stats,
