+++ NAMESPACE
@@ -18,7 +18,9 @@ export(list_identifiers)
 export(list_metadataformats)
 export(list_records)
 export(list_sets)
+export(load_providers)
 export(type_sum)
+export(update_providers)
 importFrom(httr,GET)
 importFrom(httr,content)
 importFrom(httr,stop_for_status)
