### 2015-08-25 1db69f5
added tests for providers data, added man file for providers dataset
