### 2015-08-25 ef4f6f7
fix missing param in update_proivders
