### 2015-08-25 7e76774
cran badges for readme
