+++ README.Rmd
@@ -12,6 +12,8 @@ knitr::opts_chunk$set(
 
 [![Build Status](https://travis-ci.org/sckott/oai.svg?branch=master)](https://travis-ci.org/sckott/oai)
 [![codecov.io](http://codecov.io/github/sckott/oai/coverage.svg?branch=master)](http://codecov.io/github/sckott/oai?branch=master)
+[![rstudio mirror downloads](http://cranlogs.r-pkg.org/badges/oai?color=2ED968)](https://github.com/metacran/cranlogs.app)
+[![cran version](http://www.r-pkg.org/badges/version/oai)](https://cran.rstudio.com/web/packages/oai)
 
 `oai` is an R client to work with OAI-PMH services.
 
