### 2015-08-25 6cf75da
added more tests, fix #1
