+++ README.Rmd
@@ -1,5 +1,5 @@
 oai
-=======
+===
 
 ```{r echo=FALSE}
 knitr::opts_chunk$set(
