### 2015-08-25 3860bcc
added get_records test
