### 2015-08-26 bffeb56
readme update with more in depth pkg and oai-pmh description
