### 2015-08-27 f1705b7
allow most fxns to give back list, df, or raw text fix #7
