+++ R/list_records.R
@@ -13,26 +13,46 @@
 #' @param token (character) a token previously provided by the server to resume a request
 #'     where it last left off. 50 is max number of records returned. We will
 #'     loop for you internally to get all the records you asked for.
+#' @param as (character) What to return. One of "df" (for data.frame; default),
+#'     "list", or "raw" (raw text)
 #' @param ... Curl options passed on to \code{\link[httr]{GET}}
 #' @examples \dontrun{
+#' # By default you get back a single data.frame
 #' list_records(from = '2011-05-01T', until = '2011-08-15T')
 #' list_records(from = '2011-05-01T', until = '2011-07-15T')
 #' list_records(from = '2011-06-01T', until = '2011-07-01T')
+#'
+#' # Get a list
+#' list_records(from = '2011-06-01T', until = '2011-07-01T', as = "list")
+#'
+#' # Get raw text
+#' list_records(from = '2011-06-01T', until = '2011-07-01T', as = "raw")
+#' list_records(from = '2011-06-01T', until = '2011-07-30T', as = "raw")
 #'
 #' # use curl options
 #' library("httr")
 #' list_records(from = '2011-05-01T', until = '2011-07-15T', config=verbose())
 #' }
 list_records <- function(url = "http://oai.datacite.org/oai", prefix = "oai_dc", from = NULL,
-                         until = NULL, set = NULL, token = NULL, ...) {
+  until = NULL, set = NULL, token = NULL, as = "df", ...) {
 
   check_url(url)
   args <- sc(list(verb = "ListRecords", metadataPrefix = prefix, from = from,
                   until = until, set = set, token = token))
-  out <- while_oai(url, args, token, ...)
-  structure(rbind_fill(out),
-            class = c("oai_df", "data.frame"),
-            type = "ListRecords")
+  out <- while_oai(url, args, token, as, ...)
+  oai_give(out, as, "ListRecords")
+}
+
+oai_give <- function(x, as, type) {
+  switch(as,
+         df = {
+           structure(rbind_fill(x),
+                     class = c("oai_df", "data.frame"),
+                     type = type)
+         },
+         list = x,
+         raw = x
+  )
 }
 
 check_url <- function(x) {
@@ -58,36 +78,44 @@ print.oai_df <- function(x, ..., n = 10) {
   trunc_mat(x, n = n)
 }
 
-get_data <- function(x) {
+get_data <- function(x, as = "df") {
   sc(lapply(x, function(z) {
     if (xml2::xml_name(z) != "resumptionToken") {
       tmp <- xml2::xml_children(z)
-      hd <- get_headers(tmp[[1]])
-      met <- get_metadata(tmp)
-      if (!is.null(met)) {
-        data.frame(hd, met, stringsAsFactors = FALSE)
-      } else {
-        hd
-      }
+      hd <- get_headers(tmp[[1]], as = as)
+      met <- get_metadata(tmp, as = as)
+      switch(as,
+             df = {
+               if (!is.null(met)) {
+                 data.frame(hd, met, stringsAsFactors = FALSE)
+               } else {
+                 hd
+               }
+             },
+             list = list(headers = hd, metadata = met),
+             raw = z
+      )
     }
   }))
 }
 
-get_headers <- function(x) {
-  rbind_df(lapply(xml2::xml_children(x), function(w) {
+get_headers <- function(m, as = "df") {
+  tmpm <- lapply(xml2::xml_children(m), function(w) {
     as.list(setNames(xml2::xml_text(w), xml2::xml_name(w)))
-  }))
+  })
[…]
