### 2015-08-27 702bad7
remove cran link from readme for now to make check happy
