### 2015-08-27 6cbe53c
better pkg level man file, fix #8
