### 2015-09-10 e9a6963
added news and conduct files, bumped to v0.1, added cran ver badge
