+++ NEWS.md
@@ -1,0 +1,4 @@
+oai 0.1.0
+=========
+
+* Released to CRAN.
