### 2015-09-10 cb419a3
remove cran ver link that check doesnt like
