### 2015-09-11 6833682
added cran version badge to readme, bumped to 99 version
