### 2015-09-21 dcd7541
Developing result dumpers
