### 2015-09-24 403027c
Written xml_to_txt dumper
Writes to files names defaulting to current timestamp.
