+++ R/dumpers.R
@@ -9,14 +9,11 @@
 #' @param args
 #' @param dumper_args a list of additional arguments accepted by the dumper
 
-dump_xml_to_txt <- function(res, args,
-                            file=ifelse(onefile, "oaidump.xml", "oaidump%04d.xml"),
-                            onefile=TRUE
-  if(!onefile) dumper_call_count <<- 1
-
-
-
-
+dump_xml_to_txt <- function(res, args, file=NULL, ...) {
+  if(is.null(file))
+    file <- paste0(chartr(": ", "-_", Sys.time()), ".xml")
+  cat(res, file=file)
+  cat("Dumped to file ", file, "\n", sep="")
 }
 
 
