### 2015-09-25 59b63cc
Alternative handle_errors2
