### 2015-09-26 040de54
Read XML safely with tests
Catch errors triggered by illegal characters: Remove the characters and
throw a warning.
