### 2015-09-28 bb3d7d1
In while_oai `as` determines what the dumper gets
