### 2015-09-28 09d2217
Enabling result dumpers
