+++ R/dumpers.R
@@ -1,17 +1,160 @@
 #' Result dumpers
 #'
-#' Result dumpers are an functions allowing to handle the results from OAI-PMH
+#' Result dumpers are functions allowing to handle the results from OAI-PMH
 #' service "on the fly". Handling can include processing, writing to files,
 #' databases etc. This package implements a couple of exemplary dumpers. See
 #' Details for how to write your own.
 #'
-#' @param res
-#' @param args
-#' @param dumper_args a list of additional arguments accepted by the dumper
+#' A dumper is a function that needs to accept at least the arguments:
+#' \code{res}, \code{args}, \code{as}.
+#'
+#' @param res results, depends on \code{as}
+#' @param args list, query arguments
+#' @param as character, type of result to return
+#' @param ... arguments passed to other functions
+#'
+#' @return
+#' Dumpers should return \code{NULL} or a value that will be collected and
+#' returned by the function calling the dumper (e.g.
+#' \code{\link{list_records}}).
+#'
+#' @aliases dumpers
+#' @name dumpers
+NULL
 
-dump_xml_to_txt <- function(res, args, file=NULL, ...) {
-  if(is.null(file))
-    file <- paste0(chartr(": ", "-_", Sys.time()), ".xml")
-  cat(res, file=file)
-  cat("Dumped to file ", file, "\n", sep="")
+
+
+#' @rdname dumpers
+#'
+#' @details
+#'
+#' `dump_raw_to_txt` writes raw XML to text files. It requires \code{as=="raw"}.
+#' File names are created with \code{\link{Sys.time}} with ":" and spaces
+#' replaced with "-" and "_" respectively. \code{file_prefix} is used when
+#' provided.
+#'
+#' @param file_prefix character or \code{NULL} a file name prefix prepended to
+#'   the target file name with \code{\link{file.path}}
+#'
+#' @return `dump_raw_to_txt` returns the name of the created file
+#'
+#' @export
+dump_raw_to_txt <- function(res, args, as, file_prefix=NULL, ...) {
+  stopifnot(as == "raw")
+  filename <- paste0(chartr(": ", "-_", Sys.time()), ".xml")
+  if( !is.null(file_prefix) ) {
+    filename <- file.path(file_prefix, filename)
+  }
+  cat( as.character(res), file=filename, ... )
+  filename
+}
+
+
+
+
+#' @rdname dumpers
+#'
+#' @details
+#'
+#' `dump_xml_to_db` writes raw XMLs to a single text column in a database.
+#' Requires \code{as == "raw"}.
+#'
+#' @param dbcon DBI-compliant database connection
+#' @param table_name character, name of the table to write into
+#' @param field_name character, name of the field to write into
+#'
+#' @return `dump_xml_to_db` returns \code{NULL}
+#'
+#' @export
+dump_raw_to_db <- function(res, args, as, dbcon, table_name, field_name, ...) {
+  stopifnot( as == "raw" )
+  dframe <- data.frame(x=as.character(res), stringsAsFactors = FALSE)
+  names(dframe) <- field_name
+  DBI::dbWriteTable(con=dbcon, name=table_name, value=dframe, row.names=FALSE, append=TRUE, ...)
+  invisible(NULL)
+}
+
+
+
+
+
+#' @rdname dumpers
+#'
+#' @details
+#'
+#' `dump_to_rds` saves results in an `.rds` file via \code{\link{saveRDS}}. Type
+#' of object being saved is determined by \code{as} argument. File names are
+#' created with \code{\link{Sys.time}} with ":" and spaces replaced with "-" and
[…]
