+++ man/dumpers.Rd
@@ -1,0 +1,68 @@
+% Generated by roxygen2 (4.1.1): do not edit by hand
+% Please edit documentation in R/dumpers.R
+\name{dumpers}
+\alias{dump_raw_to_db}
+\alias{dump_raw_to_txt}
+\alias{dump_to_rds}
+\alias{dumpers}
+\title{Result dumpers}
+\usage{
+dump_raw_to_txt(res, args, as, file_prefix = NULL, ...)
+
+dump_raw_to_db(res, args, as, dbcon, table_name, field_name, ...)
+
+dump_to_rds(res, args, as, file_prefix = NULL, ...)
+}
+\arguments{
+\item{res}{results, depends on \code{as}}
+
+\item{args}{list, query arguments}
+
+\item{as}{character, type of result to return}
+
+\item{file_prefix}{character or \code{NULL} a file name prefix prepended to
+  the target file name with \code{\link{file.path}}}
+
+\item{...}{arguments passed to other functions}
+
+\item{dbcon}{DBI-compliant database connection}
+
+\item{table_name}{character, name of the table to write into}
+
+\item{field_name}{character, name of the field to write into}
+}
+\value{
+Dumpers should return \code{NULL} or a value that will be collected and
+returned by the function calling the dumper (e.g.
+\code{\link{list_records}}).
+
+`dump_raw_to_txt` returns the name of the created file
+
+`dump_xml_to_db` returns \code{NULL}
+
+`dump_to_rds` returns the name of the created file.
+}
+\description{
+Result dumpers are functions allowing to handle the results from OAI-PMH
+service "on the fly". Handling can include processing, writing to files,
+databases etc. This package implements a couple of exemplary dumpers. See
+Details for how to write your own.
+}
+\details{
+A dumper is a function that needs to accept at least the arguments:
+\code{res}, \code{args}, \code{as}.
+
+`dump_raw_to_txt` writes raw XML to text files. It requires \code{as=="raw"}.
+File names are created with \code{\link{Sys.time}} with ":" and spaces
+replaced with "-" and "_" respectively. \code{file_prefix} is used when
+provided.
+
+`dump_xml_to_db` writes raw XMLs to a single text column in a database.
+Requires \code{as == "raw"}.
+
+`dump_to_rds` saves results in an `.rds` file via \code{\link{saveRDS}}. Type
+of object being saved is determined by \code{as} argument. File names are
+created with \code{\link{Sys.time}} with ":" and spaces replaced with "-" and
+"_" respectively. \code{file_prefix} is used when provided.
+}
+
