+++ NAMESPACE
@@ -12,6 +12,9 @@ S3method(type_sum,logical)
 S3method(type_sum,matrix)
 S3method(type_sum,numeric)
 export(count_identifiers)
+export(dump_raw_to_db)
+export(dump_raw_to_txt)
+export(dump_to_rds)
 export(get_records)
 export(id)
 export(list_identifiers)
