### 2015-09-28 bf9d0bd
Added dumper tests
