+++ tests/testthat/test-dumpers.R
@@ -1,0 +1,168 @@
+
+# valid_dumper ------------------------------------------------------------
+
+context("Dumper validity checker")
+
+dummy_dumper <- function(res, args, as, x=1, ...) {
+  c(list(res=res, args=args, as=as, x=x), list(...))
+}
+
+
+test_that("supplied all obligatory arguments works", {
+  expect_true( valid_dumper( function(res, args, as) {}, NULL))
+} )
+
+test_that("missing obligatory argument throws error", {
+  expect_error( valid_dumper(  function(res, args) {}, NULL ) )
+  expect_error( valid_dumper(  function(res, as) {}, NULL ) )
+  expect_error( valid_dumper(  function(args, as) {}, NULL ) )
+} )
+
+test_that("supplying additional arguments unaccepted by dumper", {
+  expect_error( valid_dumper( function(res, args, as) {}, list(a=1) ) )
+  expect_true( valid_dumper( function(res, args, as, ...) {}, list(a=1) ) )
+} )
+
+test_that("missing necessary extra argument", {
+  expect_error( valid_dumper(
+    function(res, args, as, x) {},
+    NULL
+  ) )
+  expect_error( valid_dumper(
+    function(res, args, as, x, ...) {},
+    list(a=1)
+  ) )
+} )
+
+test_that("user cannot provide res/args/as arguments", {
+  expect_error( valid_dumper( function(res, args, as) {}, list(res=1) ) )
+  expect_error( valid_dumper( function(res, args, as) {}, list(args=1) ) )
+  expect_error( valid_dumper( function(res, args, as) {}, list(as=1) ) )
+} )
+
+
+test_that("user did not supply extra argument but dumper has a default", {
+  expect_true( valid_dumper(dummy_dumper, NULL) )
+} )
+
+
+
+
+# dump_raw_to_txt ---------------------------------------------------------
+
+
+
+context("Testing text file dumper")
+
+test_that("list_identifiers saves raw XML to text files", {
+  fnames <- list_identifiers(from=as.character(Sys.Date()-1), as="raw",
+                             dumper=dump_raw_to_txt)
+  expect_true(is.character(fnames))
+  expect_true(all(sapply(fnames, file.exists)))
+
+  xmls <- lapply(fnames, xml2::read_xml)
+  expect_true( all(sapply(xmls, inherits, "xml_document")) )
+
+  unlink(fnames)
+} )
+
+
+test_that("list_records saves raw XML to text files", {
+  fnames <- list_records(from = '2011-06-01T', until = '2011-07-01T',
+                         as = "raw",
+                         dumper=dump_raw_to_txt )
+  expect_true(is.character(fnames))
+  expect_true(all(sapply(fnames, file.exists)))
+
+  xmls <- lapply(fnames, xml2::read_xml)
+  expect_true( all(sapply(xmls, inherits, "xml_document")) )
+
+  unlink(fnames)
+} )
+
+
+
+
+
+# dump_to_rds -------------------------------------------------------------
+
+
+
+
+context("Testing RDS file dumper")
+
+test_that("list_identifiers saves raw XML to RDS files", {
+  fnames <- list_identifiers(from=as.character(Sys.Date()-1), as="raw",
+                             dumper=dump_to_rds)
+  expect_true(is.character(fnames))
+  expect_true(all(sapply(fnames, file.exists)))
+
+  # read and parse
+  xmls <- lapply(fnames, function(fn) xml2::read_xml(readRDS(fn)))
+  expect_true( all(sapply(xmls, inherits, "xml_document")) )
+
+  unlink(fnames)
+} )
+
+
+test_that("list_records saves raw XML to RDS files", {
[…]
