### 2015-09-28 d216bc6
Dumper documentation fixes
