### 2015-09-28 935b2ac
Cleanups in dumpers
