### 2015-09-28 e2cc324
Using new handle_errors in while_oai and get_record
