+++ R/list_records.R
@@ -65,17 +65,11 @@ is.url <- function(x, ...){
   grepl("https?://", x)
 }
 
-handle_errors <- function(x) {
-  nms <- xml2::xml_name(xml2::xml_children(xml2::read_xml(x)))
-  if ("error" %in% nms) {
-    stop(xml2::xml_text(xml2::xml_children(xml2::read_xml(x))[[3]]), call. = FALSE)
-  }
-}
 
 # Look for OAI-PMH exceptions
 # https://www.openarchives.org/OAI/openarchivesprotocol.html#ErrorConditions
 # xml = parsed xml
-handle_errors2 <- function(xml) {
+handle_errors <- function(xml) {
   nodeset <- xml2::xml_find_all(xml, ".//*[local-name()='error']")
   if( length(nodeset) > 0 ) {
     msgs <- sapply(nodeset, function(n)
