+++ R/get_record.R
@@ -33,11 +33,11 @@ each_record <- function(identifier, url, prefix, as, ...) {
   res <- GET(url, query = args, ...)
   stop_for_status(res)
   tt <- content(res, "text")
-  handle_errors(tt)
+  xml_orig <- xml2::read_xml(tt)
+  handle_errors(xml_orig)
   if (as == "raw") {
     tt
   } else {
-    xml_orig <- xml2::read_xml(tt)
     xml <- xml2::xml_children(xml2::xml_children(xml_orig)[[3]])
     get_data(xml, as = as)
   }
