### 2015-09-28 41acf88
Fixed test of list_identifiers to lookup yesterday data
Testing today might lead to no data.
