+++ DESCRIPTION
@@ -7,10 +7,11 @@ Description: A general purpose client to work with any 'OAI-PMH'
     Functions are provided to work with the 'OAI-PMH' verbs: 'GetRecord',
     'Identify', 'ListIdentifiers', 'ListMetadataFormats', 'ListRecords', and
     'ListSets'.
-Version: 0.1.0.9000
+Version: 0.1.0.9400
 License: MIT + file LICENSE
-Authors@R: c(person("Scott", "Chamberlain", role = c("aut", "cre"),
-    email = "myrmecocystus@gmail.com"))
+Authors@R: c(
+    person("Scott", "Chamberlain", role = c("aut", "cre"), email = "myrmecocystus@gmail.com"),
+    person("Michal", "Bojanowski", role = "aut"))
 URL: https://github.com/sckott/oai
 BugReports: https://github.com/sckott/oai/issues
 VignetteBuilder: knitr
