### 2015-09-28 0a0dacd
Updated functions to use read_xml_safely
Updates: count_identifiers, get_record, id, list_metadataformats, while_oai
