### 2015-09-28 38e5dd7
fix to id() egs as url param has no default also, state that url is required in man file
