### 2015-09-29 e692845
Fixed read_xml_safely tests
Now comparing as.character(xml).
