+++ tests/testthat/test-read_xml_safely.R
@@ -6,13 +6,15 @@ inputs <- c(
 )
 
 test_that("proper XML is parsed correctly", {
-  expect_equal( read_xml_safely(inputs[1]), xml2::read_xml(inputs[1]) )
+  expect_equal( as.character(read_xml_safely(inputs[1])),
+                as.character(xml2::read_xml(inputs[1])) )
 } )
 
 
 test_that("RESTART: improper characters are translated", {
   expect_warning( res <- read_xml_safely(inputs[2]) )
-  expect_equal(res, xml2::read_xml( gsub("\u0019", "", inputs[2]) ))
+  expect_equal( as.character(res),
+                as.character(xml2::read_xml( gsub("\u0019", "", inputs[2]))) )
 } )
 
 
