### 2015-10-01 32bbda2
Fixes for dumpers
Files names are generated using tempfile().
Added more documentation.
