+++ man/dumpers.Rd
@@ -7,11 +7,13 @@
 \alias{dumpers}
 \title{Result dumpers}
 \usage{
-dump_raw_to_txt(res, args, as, file_prefix = NULL, ...)
+dump_raw_to_txt(res, args, as, file_pattern = "oaidump", file_dir = ".",
+  file_ext = ".xml")
+
+dump_to_rds(res, args, as, file_pattern = "oaidump", file_dir = ".",
+  file_ext = ".rds")
 
 dump_raw_to_db(res, args, as, dbcon, table_name, field_name, ...)
-
-dump_to_rds(res, args, as, file_prefix = NULL, ...)
 }
 \arguments{
 \item{res}{results, depends on \code{as}, not to be specified by the user}
@@ -21,50 +23,83 @@ dump_to_rds(res, args, as, file_prefix = NULL, ...)
 \item{as}{character, type of result to return, not to be specified by the
 user}
 
-\item{file_prefix}{character or \code{NULL} a file name prefix prepended to
-  the target file name with \code{\link{file.path}}}
+\item{file_pattern,file_dir,file_ext}{character respectively: initial part of
+  the file name, directory name, and file extension used to create file
+  names. These arguments are passed to \code{\link{tempfile}} arguments
+  \code{pattern}, \code{tmpdir}, and \code{fileext} respectively.}
 
-\item{...}{arguments passed to other functions}
+\item{dbcon}{\pkg{DBI}-compliant database connection}
 
-\item{dbcon}{DBI-compliant database connection}
+\item{table_name}{character, name of the database table to write into}
 
-\item{table_name}{character, name of the table to write into}
+\item{field_name}{character, name of the field in database table to write
+  into}
 
-\item{field_name}{character, name of the field to write into}
+\item{...}{arguments passed to/from other functions}
 }
 \value{
 Dumpers should return \code{NULL} or a value that will be collected
-and returned by the function calling the dumper (e.g.
-\code{\link{list_records}}).
+  and returned by the function calling the dumper (e.g.
+  \code{\link{list_records}}).
 
-`dump_raw_to_txt` returns the name of the created file
+\code{dump_raw_to_txt} returns the name of the created file the
+  results have beed written into.
 
-`dump_xml_to_db` returns \code{NULL}
+`dump_to_rds` returns a vector of names of created files.
 
-`dump_to_rds` returns the name of the created file.
+\code{dump_xml_to_db} returns \code{NULL}
 }
 \description{
-Result dumpers are functions allowing to handle the results from OAI-PMH
-service "on the fly". Handling can include processing, writing to files,
-databases etc. This package implements a couple of exemplary dumpers. See
-Details for how to write your own.
+Result dumpers are functions allowing to handle the chunks of results from
+OAI-PMH service "on the fly". Handling can include processing, writing to
+files, databases etc.
 }
 \details{
-A dumper is a function that needs to accept at least the arguments:
+Often the result of a request to a OAI-PMH service are so large that it is
+split into chunks that need to be requested separately using
+\code{resumptionToken}. By default functions like
+\code{\link{list_identifiers}} or \code{\link{list_records}} request these
+chunks under the hood and return all concatenated in a single R object. It is
+convenient but insufficient when dealing with large result sets that might
+not fit into RAM. A result dumper is a function that is called on each result
+chunk. Chunks can be written to files or databases with initial
+pre-processing, and so on.
+
+A result dumper needs to be function that accepts at least the arguments:
 \code{res}, \code{args}, \code{as}. They will get values by the enclosing
-function internally.
+function internally. There may be additional arguments, including \code{...}.
[…]
