### 2015-10-01 496f9a5
Updated dumper tests
