+++ tests/testthat/test-dumpers.R
@@ -55,10 +55,9 @@ test_that("user did not supply extra argument but dumper has a default", {
 context("Testing text file dumper")
 
 test_that("list_identifiers saves raw XML to text files", {
-  tdir <- tempdir()
   fnames <- list_identifiers(from=as.character(Sys.Date()-1), as="raw",
                              dumper=dump_raw_to_txt,
-                             dumper_args=list(file_prefix=tdir))
+                             dumper_args=list(file_dir=tempdir()))
   expect_true(is.character(fnames))
   expect_true(all(sapply(fnames, file.exists)))
 
@@ -72,7 +71,8 @@ test_that("list_identifiers saves raw XML to text files", {
 test_that("list_records saves raw XML to text files", {
   fnames <- list_records(from = '2011-06-01T', until = '2011-07-01T',
                          as = "raw",
-                         dumper=dump_raw_to_txt )
+                         dumper=dump_raw_to_txt,
+                         dumper_args=list(file_dir=tempdir()))
   expect_true(is.character(fnames))
   expect_true(all(sapply(fnames, file.exists)))
 
@@ -95,7 +95,8 @@ context("Testing RDS file dumper")
 
 test_that("list_identifiers saves raw XML to RDS files", {
   fnames <- list_identifiers(from=as.character(Sys.Date()-1), as="raw",
-                             dumper=dump_to_rds)
+                             dumper=dump_to_rds,
+                             dumper_args=list(file_dir=tempdir()))
   expect_true(is.character(fnames))
   expect_true(all(sapply(fnames, file.exists)))
 
@@ -110,7 +111,8 @@ test_that("list_identifiers saves raw XML to RDS files", {
 test_that("list_records saves raw XML to RDS files", {
   fnames <- list_records(from = '2011-06-01T', until = '2011-07-01T',
                          as = "raw",
-                         dumper=dump_to_rds )
+                         dumper=dump_to_rds,
+                         dumper_args=list(file_dir=tempdir()))
   expect_true(is.character(fnames))
   expect_true(all(sapply(fnames, file.exists)))
 
