### 2015-10-01 18a4f06
while_oai checks if supplied dumper is valid
