+++ R/while_oai.R
@@ -41,6 +41,7 @@ while_oai <- function(url, args, token, as, dumper=NULL, dumper_args=NULL, ...) 
     if(is.null(dumper)) {
       out[[iter]] <- res
     } else {
+      valid_dumper(dumper, dumper_args)
       dumper_res <- do.call("dumper", c(list(res=res, args=args2, as=as), dumper_args))
       if(!is.null(dumper_res))
         out[[iter]] <- dumper_res
