### 2015-10-01 0dc9c13
Dumper documentation fixes
