+++ R/dumpers.R
@@ -11,8 +11,8 @@
 #' chunks under the hood and return all concatenated in a single R object. It is
 #' convenient but insufficient when dealing with large result sets that might
 #' not fit into RAM. A result dumper is a function that is called on each result
-#' chunk. Chunks can be written to files or databases with initial
-#' pre-processing, and so on.
+#' chunk. Dumper functions can write chunks to files or databases, include initial
+#' pre-processing or extraction, and so on.
 #'
 #' A result dumper needs to be function that accepts at least the arguments:
 #' \code{res}, \code{args}, \code{as}. They will get values by the enclosing
@@ -21,12 +21,18 @@
 #' be collected and returned by the function calling the dumper (e.g.
 #' \code{\link{list_records}}).
 #'
-#' To use a dumper you need to:
+#' Currently result dumpers can be used with functions:
+#' \code{\link{list_identifiers}},
+#' \code{\link{list_records}}, and
+#' \code{\link{list_sets}}.
+#' To use a dumper with one of these functions you need to:
 #' \enumerate{
-#' \item{Pass it as an argument \code{dumper}}
-#' \item{Addtional arguments to a dumper might be passed as a list to \code{dumper_args} argument}
+#' \item{Pass it as an additional argument \code{dumper}}
+#' \item{Pass optional addtional arguments to the dumper function in a list as the \code{dumper_args} argument}
 #' }
-#' See Examples.
+#' See Examples. Below we provide more details on the dumpers currently implemented.
+#'
+#'
 #'
 #' @param res results, depends on \code{as}, not to be specified by the user
 #' @param args list, query arguments, not to be specified by the user
@@ -35,11 +41,15 @@
 #' @param ... arguments passed to/from other functions
 #'
 #' @return Dumpers should return \code{NULL} or a value that will be collected
-#'   and returned by the function calling the dumper (e.g.
-#'   \code{\link{list_records}}).
+#'   and returned by the function using the dumper.
 #'
 #' @references OAI-PMH specification
 #'   \url{https://www.openarchives.org/OAI/openarchivesprotocol.html}
+#'
+#' @seealso Functions supporting the dumpers:
+#' \code{\link{list_identifiers}},
+#' \code{\link{list_sets}}, and
+#' \code{\link{list_records}}
 #'
 #' @aliases dumpers
 #' @name dumpers
@@ -53,7 +63,7 @@ NULL
 #'
 #' \code{dump_raw_to_txt} writes raw XML to text files. It requires
 #' \code{as=="raw"}. File names are created using \code{\link{tempfile}}. By
-#' default they are written in the current directory and have a format
+#' default they are written in the current working directory and have a format
 #' \code{oaidump*.xml} where \code{*} is a random string in hex.
 #'
 #' @param file_pattern,file_dir,file_ext character respectively: initial part of
@@ -61,8 +71,7 @@ NULL
 #'   names. These arguments are passed to \code{\link{tempfile}} arguments
 #'   \code{pattern}, \code{tmpdir}, and \code{fileext} respectively.
 #'
-#' @return \code{dump_raw_to_txt} returns the name of the created file the
-#'   results have beed written into.
+#' @return \code{dump_raw_to_txt} returns the name of the created file.
 #'
 #' @export
 dump_raw_to_txt <- function(res, args, as, file_pattern="oaidump", file_dir=".",
@@ -79,11 +88,11 @@ dump_raw_to_txt <- function(res, args, as, file_pattern="oaidump", file_dir=".",
 #' @details
 #'
 #' \code{dump_to_rds} saves results in an \code{.rds} file via \code{\link{saveRDS}}.
-#' Type of object being saved is determined by \code{as} argument. File names
[…]
