### 2015-10-01 e7221f6
Added some dumper examples
