+++ man/dumpers.Rd
@@ -100,6 +100,46 @@ package \pkg{DBI}. As such, it can connect to any database supported by
 exist, it is created. If it does, the records are appended. Any additional
 arguments are passed to \code{\link[DBI]{dbWriteTable}}.
 }
+\examples{
+\dontrun{
+
+### Dumping raw XML to text files
+
+# This will write a set of XML files to a temporary directory
+fnames <- list_identifiers(from="2015-09-01",
+                           until="2015-09-01",
+                           as="raw",
+                           dumper=dump_raw_to_txt,
+                           dumper_args=list(file_dir=tempdir()))
+# vector of file names created
+str(fnames)
+all( file.exists(fnames) )
+# clean-up
+unlink(fnames)
+
+
+### Dumping raw XML to a database
+
+# Connect to in-memory SQLite database
+con <- DBI::dbConnect(RSQLite::SQLite(), dbname=":memory:")
+# Harvest and dump the results into field "bar" of table "foo"
+list_identifiers(from="2015-09-01",
+                 until="2015-09-01",
+                 as="raw",
+                 dumper=dump_raw_to_db,
+                 dumper_args=list(dbcon=con,
+                                  table_name="foo",
+                                  field_name="bar") )
+# Count records, should be 101
+DBI::dbGetQuery(con, "SELECT count(*) as no_records FROM foo")
+
+dbDisconnect(con)
+
+
+
+
+}
+}
 \references{
 OAI-PMH specification
   \url{https://www.openarchives.org/OAI/openarchivesprotocol.html}
