### 2015-10-01 65be0b1
Dumper examples in man-roxygen
