### 2015-10-02 8746e3a
Catching read_xml errors
