### 2015-10-02 1adbf80
fixes to list_records for resumptionToken #18 also made same fix to list_identifiers fxn bumped 9000 version
