+++ man/list_records.Rd
@@ -47,6 +47,9 @@ list_records(from = '2011-06-01T', until = '2011-07-01T', as = "list")
 list_records(from = '2011-06-01T', until = '2011-07-01T', as = "raw")
 list_records(from = '2011-06-01T', until = '2011-07-30T', as = "raw")
 
+# Use a resumption token
+list_records(token = "1443799900201,2015-09-01T00:00:00Z,2015-10-01T23:59:59Z,50,null,oai_dc")
+
 # use curl options
 library("httr")
 list_records(from = '2011-05-01T', until = '2011-07-15T', config=verbose())
