### 2015-10-03 bbdd268
Fixed from/until dates
