+++ tests/testthat/test-dumpers.R
@@ -55,7 +55,7 @@ test_that("user did not supply extra argument but dumper has a default", {
 context("Testing text file dumper")
 
 test_that("list_identifiers saves raw XML to text files", {
-  fnames <- list_identifiers(from=as.character(Sys.Date()-1), as="raw",
+  fnames <- list_identifiers(from = '2014-06-01T', until = '2014-06-01T', as="raw",
                              dumper=dump_raw_to_txt,
                              dumper_args=list(file_dir=tempdir()))
   expect_true(is.character(fnames))
@@ -69,7 +69,7 @@ test_that("list_identifiers saves raw XML to text files", {
 
 
 test_that("list_records saves raw XML to text files", {
-  fnames <- list_records(from = '2011-06-01T', until = '2011-07-01T',
+  fnames <- list_records(from = '2014-06-01T', until = '2014-06-01T',
                          as = "raw",
                          dumper=dump_raw_to_txt,
                          dumper_args=list(file_dir=tempdir()))
@@ -94,7 +94,7 @@ test_that("list_records saves raw XML to text files", {
 context("Testing RDS file dumper")
 
 test_that("list_identifiers saves raw XML to RDS files", {
-  fnames <- list_identifiers(from=as.character(Sys.Date()-1), as="raw",
+  fnames <- list_identifiers(from="2014-06-01T", to="2014-06-01T", as="raw",
                              dumper=dump_to_rds,
                              dumper_args=list(file_dir=tempdir()))
   expect_true(is.character(fnames))
@@ -109,7 +109,7 @@ test_that("list_identifiers saves raw XML to RDS files", {
 
 
 test_that("list_records saves raw XML to RDS files", {
-  fnames <- list_records(from = '2011-06-01T', until = '2011-07-01T',
+  fnames <- list_records(from = '2014-06-01T', until = '2014-06-01T',
                          as = "raw",
                          dumper=dump_to_rds,
                          dumper_args=list(file_dir=tempdir()))
@@ -134,7 +134,7 @@ context("Testing raw_to_db dumper with SQLite")
 
 test_that("list_identifiers dumps results to SQLite", {
   con <- DBI::dbConnect(RSQLite::SQLite(), dbname=":memory:")
-  dumprval <- list_identifiers(from=as.character(Sys.Date()-1), as="raw",
+  dumprval <- list_identifiers(from = '2014-06-01T', until = '2014-06-01T', as="raw",
                              dumper=dump_raw_to_db,
                              dumper_args=list(dbcon=con, table_name="foo",
                                               field_name="bar") )
@@ -154,7 +154,7 @@ test_that("list_identifiers dumps results to SQLite", {
 
 test_that("list_records dumps results to SQLite", {
   con <- DBI::dbConnect(RSQLite::SQLite(), dbname=":memory:")
-  dumprval <- list_records(from = '2011-06-01T', until = '2011-07-01T',
+  dumprval <- list_records(from = '2014-06-01T', until = '2014-06-01T',
                            as="raw",
                            dumper=dump_raw_to_db,
                            dumper_args=list(dbcon=con, table_name="foo",
