### 2015-10-03 7b25d95
Faster dumper tests
