### 2015-10-03 793e61a
require latest httr to avoid problems with old httr version
