### 2015-10-05 7372259
Handle errors in separate file
