### 2015-10-05 da28261
Even better OAI-PMH error handling
Properly store error codes and messages, which will allow, potentially, to handle them somewhere else.
