### 2015-10-05 e1811b4
Improved tests of handle_errors
