### 2015-10-05 44c44db
Skip tests on CRAN
