### 2015-10-05 3222137
Getting resumptionToken in separate function
