### 2015-10-05 d411bac
Parsing XML safely by deleting illegal characters
