+++ R/while_oai.R
@@ -16,7 +16,7 @@ while_oai <- function(url, args, token, as, dumper=NULL, dumper_args=NULL, ...) 
     res <- GET(url, query = args2, ...)
     stop_for_status(res)
     tt <- content(res, "text")
-    xml_orig <- xml2::read_xml(tt)
+    xml_orig <- read_xml_safely(tt)
     handle_errors(xml_orig)
     tok <- get_token(xml_orig, verb=args2$verb)
     xml <- xml2::xml_children(xml2::xml_children(xml_orig)[[3]])
