### 2015-10-05 d3d658a
Extending the loop to catch XML faults
