### 2015-10-05 409c855
remove devel travis build for now
