+++ .travis.yml
@@ -28,4 +28,3 @@ env:
   matrix:
     - RVERSION=oldrel
     - RVERSION=release
-    - RVERSION=devel COVERAGE=true
