### 2015-10-05 ac1cca2
fix to check_url to handle >1 url fix #25 also, adjusted stop message in same fxn fixed tests accordingly bumped patch version
