+++ R/list_records.R
@@ -46,78 +46,3 @@ list_records <- function(url = "http://oai.datacite.org/oai", prefix = "oai_dc",
   out <- while_oai(url, args, token, as, ...)
   oai_give(out, as, "ListRecords")
 }
-
-oai_give <- function(x, as, type) {
-  switch(as,
-         df = {
-           structure(rbind_fill(x),
-                     class = c("oai_df", "data.frame"),
-                     type = type)
-         },
-         list = x,
-         raw = x
-  )
-}
-
-check_url <- function(x) {
-  if (!is.url(x)) {
-    stop("Your URL appears to not be a proper URL", call. = FALSE)
-  }
-}
-
-is.url <- function(x, ...){
-  grepl("https?://", x)
-}
-
-
-
-#' @export
-print.oai_df <- function(x, ..., n = 10) {
-  cat(sprintf("<%s> %s X %s", attr(x, "type"), NROW(x), NCOL(x)), "\n\n")
-  trunc_mat(x, n = n)
-}
-
-get_data <- function(x, as = "df") {
-  sc(lapply(x, function(z) {
-    if (xml2::xml_name(z) != "resumptionToken") {
-      tmp <- xml2::xml_children(z)
-      hd <- get_headers(tmp[[1]], as = as)
-      met <- get_metadata(tmp, as = as)
-      switch(as,
-             df = {
-               if (!is.null(met)) {
-                 data.frame(hd, met, stringsAsFactors = FALSE)
-               } else {
-                 hd
-               }
-             },
-             list = list(headers = hd, metadata = met),
-             raw = z
-      )
-    }
-  }))
-}
-
-get_headers <- function(m, as = "df") {
-  tmpm <- lapply(xml2::xml_children(m), function(w) {
-    as.list(setNames(xml2::xml_text(w), xml2::xml_name(w)))
-  })
-  switch(as, df = rbind_df(tmpm), list = unlist(tmpm, recursive = FALSE))
-}
-
-get_metadata <- function(x, as = "df") {
-  status <- unlist(xml2::xml_attrs(x))
-  if (length(status) != 0) {
-    NULL
-  } else {
-    tmp <- xml2::xml_children(xml2::xml_children(x))
-    tmpm <- lapply(tmp, function(w) {
-      as.list(setNames(xml2::xml_text(w), xml2::xml_name(w)))
-    })
-    switch(as, df = rbind_df(tmpm), list = unlist(tmpm, recursive = FALSE))
-  }
-}
-
-rbind_df <- function(x) {
-  data.frame(rbind(unlist(x)), stringsAsFactors = FALSE)
-}
