+++ tests/testthat/test-list_records.R
@@ -17,5 +17,5 @@ test_that("list_records fails well", {
                "The request includes illegal arguments")
   expect_error(list_records(from = '2011-06-01T', until = 5),
                "The request includes illegal arguments")
-  expect_error(list_records(url = 5), "Your URL appears to not be a proper URL")
+  expect_error(list_records(url = 5), "One or more of your URLs")
 })
