+++ tests/testthat/test-count_identifiers.R
@@ -53,5 +53,5 @@ test_that("count_identifiers fails well", {
   expect_error(list_sets(token = 454),
                "The value of the resumptionToken argument is invalid or expired")
   expect_error(list_sets("stuff"),
-               "Your URL appears to not be a proper URL")
+               "One or more of your URLs")
 })
