### 2015-10-05 94d1d3c
speed up tests for count_identifiers and list_identifiers #19 also while doing that moved around some helper internal fxns for better organization
