### 2015-10-06 4474317
read_xml_safely cleanups
is.na(invalid_as) leads to no replacing
