+++ R/while_oai.R
@@ -16,32 +16,41 @@ while_oai <- function(url, args, token, as, dumper=NULL, dumper_args=NULL, ...) 
     res <- GET(url, query = args2, ...)
     stop_for_status(res)
     tt <- content(res, "text")
-    xml_orig <- try( read_xml_safely(tt) )
-    if(inherits(xml_orig, "error")) {
-      warning("XML parsing failed, trying read_html")
-      html_orig <- try( xml2::read_html(tt) )
-      if( inherits(html_orig, "error") ) {
-        fname <- tempfile()
+
+    # try parsing
+    parsed <- try( read_xml_safely(tt), silent=TRUE )
+    if(inherits(parsed, "try-error")) {
+      parsed <- try( xml2::read_html(tt), silent=TRUE )
+      warning("read_xml parsing failed, but read_html succeeded")
+      if( inherits(parsed, "try-error") ) {
+        fname <- tempfile(
+          pattern=paste0("oaidump_", chartr(" :", "_-", Sys.time())),
+          tmpdir=".", fileext="xml")
         cat(tt, file=fname)
-        stop(paste0("cannot parse XML, dumping to file ", fname))
-      } else {
-        handle_errors(html_orig)
-        tok <- get_token(html_orig, verb=args2$verb, is_html=TRUE)
+        stop(paste0("cannot parse downloaded XML, dumping to file ", fname))
       }
+      is_html <- TRUE
     } else {
-      handle_errors(xml_orig)
-      tok <- get_token(xml_orig, verb=args2$verb)
+      is_html <- FALSE
     }
+
+    handle_errors(parsed)
+    tok <- get_token(parsed, verb=args2$verb, is_html=is_html)
     # `as` determines what the `dumper` gets
-    res <- if (as == "raw") {
-      tt
+    if (as == "raw") {
+      res <- tt
     } else {
-      xml <- xml2::xml_children(xml2::xml_children(xml_orig)[[3]])
-      switch(args$verb,
-             ListRecords = get_data(xml, as = as),
-             ListIdentifiers = parse_listid(xml, as = as),
-             ListSets = get_sets(xml, as = as)
-      )
+      if(is_html) {
+        warning("malformed XML - keeping raw text even though `as` is ", dQuote(as))
+        res <- tt
+      } else {
+        xml_verb <- xml2::xml_children(xml2::xml_children(parsed)[[3]])   # TODO xpath here
+        res <- switch(args$verb,
+               ListRecords = get_data(xml_verb, as = as),
+               ListIdentifiers = parse_listid(xml_verb, as = as),
+               ListSets = get_sets(xml_verb, as = as)
+        )
+      }
     }
     # Collect values returned by `dumper` if they are not NULL
     if (is.null(dumper)) {
@@ -72,11 +81,15 @@ while_oai <- function(url, args, token, as, dumper=NULL, dumper_args=NULL, ...) 
 get_token <- function(x, verb, is_html=FALSE) {
   xp <- paste0("/*[local-name()='OAI-PMH']/*[local-name()='", verb, "']/*[local-name()='",
                if(is_html) "resumptiontoken" else "resumptionToken", "']" )
-  node <- xml2::xml_find_one(x, xp)
+  node <- xml2::xml_find_all(x, xp)
   if(length(node) == 0) {
     return( list(token="") )
   }
   else {
+    if(length(node) > 1) {
+      warning("more than one match - using last")
+      node <- node[length(node)]
+    }
     return( c(
       token=xml2::xml_text(node),
       as.list(xml2::xml_attrs(node))
