### 2015-10-19 #19 @karthik
## Review of OAI

Major issues
The package works as intended and I did not find any major issues with the intended functionality.

Minor issues:

- The acronym (OAI) is never expanded even once. It would be nice for the `DESCRIPTION` file to say "Open Archives Initiative Protocol for Metadata Harvesting". It is mentioned in the README which I appreciate.
- List suggests in alphabetical order [recommended here](http://r-pkgs.had.co.nz/description.html).
- On package load:
```
The following objects are masked from 'package:dplyr':

    id, type_sum
```

Could these benefit from less conflicting names? I know it's already on CRAN now which makes it difficult to change, but avoiding conflicts with popular packages like `dplyr` is generally helpful.
- `trunc_mat` and other functions in `tbl_df.R` could use at least a one line commented description even if they are internal functions.
- Is there a reason `rbind_fill` is copied over from `plyr` into the package besides trying to minimize one more dependency? I would suggest not doing that in figure since it doesn't seem like good practice to duplicate functions that way.

### tests
All tests pass for me locally under R `3.2.2` and R devel.

_with apologies for the slow review._ :disappointed:
