### 2015-10-19 #19 @karthik
I forgot to add. No objections to moving this into the ropensci org account.
