### 2015-10-21 #19 @sckott
> The acronym (OAI) is never expanded even once. It would be nice for the `DESCRIPTION` file to say "Open Archives Initiative Protocol for Metadata Harvesting". It is mentioned in the README which I appreciate.

will do

> List suggests in alphabetical order [recommended here](http://r-pkgs.had.co.nz/description.html).

Will do

> On package load: Could these benefit from less conflicting names? I know it's already on CRAN now which makes it difficult to change, but avoiding conflicts with popular packages like `dplyr` is generally helpful.

yeah, can change names

> `trunc_mat` and other functions in `tbl_df.R` could use at least a one line commented description even if they are internal functions.

will do

> Is there a reason `rbind_fill` is copied over from `plyr` into the package besides trying to minimize one more dependency? I would suggest not doing that in figure since it doesn't seem like good practice to duplicate functions that way.

To reduce dependencies. But I'll just import plyr
