+++ DESCRIPTION
@@ -2,12 +2,13 @@ Package: oai
 Type: Package
 Title: General Purpose 'Oai-PMH' Services Client
 Description: A general purpose client to work with any 'OAI-PMH'
-    service. The 'OAI-PMH' protocol is described at
+    (Open Archives Initiative Protocol for Metadata Harvesting) service.
+    The 'OAI-PMH' protocol is described at
     http://www.openarchives.org/OAI/openarchivesprotocol.html.
     Functions are provided to work with the 'OAI-PMH' verbs: 'GetRecord',
     'Identify', 'ListIdentifiers', 'ListMetadataFormats', 'ListRecords', and
     'ListSets'.
-Version: 0.1.1.9000
+Version: 0.1.1.9100
 License: MIT + file LICENSE
 Authors@R: c(
     person("Scott", "Chamberlain", role = c("aut", "cre"), email = "myrmecocystus@gmail.com"),
