### 2015-10-21 d32f54b
import plyr::rbind.fill, bumped 99 version fix #32
