+++ R/rbind_fill.R
@@ -1,167 +1,0 @@
-rbind_fill <- function (...) {
-  dfs <- list(...)
-  if (length(dfs) == 0)
-    return()
-  if (is.list(dfs[[1]]) && !is.data.frame(dfs[[1]])) {
-    dfs <- dfs[[1]]
-  }
-  dfs <- sc(dfs)
-  if (length(dfs) == 0)
-    return()
-  if (length(dfs) == 1)
-    return(dfs[[1]])
-  is_df <- vapply(dfs, is.data.frame, logical(1))
-  if (any(!is_df)) {
-    stop("All inputs to rbind_fill must be data.frames",
-         call. = FALSE)
-  }
-  rows <- unlist(lapply(dfs, .row_names_info, 2L))
-  nrows <- sum(rows)
-  ot <- output_template(dfs, nrows)
-  setters <- ot$setters
-  getters <- ot$getters
-  if (length(setters) == 0) {
-    return(as.data.frame(matrix(nrow = nrows, ncol = 0)))
-  }
-  pos <- matrix(c(cumsum(rows) - rows + 1, rows), ncol = 2)
-  for (i in seq_along(rows)) {
-    rng <- seq(pos[i, 1], length = pos[i, 2])
-    df <- dfs[[i]]
-    for (var in names(df)) {
-      setters[[var]](rng, df[[var]])
-    }
-  }
-  quickdf(lapply(getters, function(x) x()))
-}
-
-output_template <- function (dfs, nrows) {
-  vars <- unique(unlist(lapply(dfs, base::names)))
-  output <- vector("list", length(vars))
-  names(output) <- vars
-  seen <- rep(FALSE, length(output))
-  names(seen) <- vars
-  for (df in dfs) {
-    matching <- intersect(names(df), vars[!seen])
-    for (var in matching) {
-      output[[var]] <- allocate_column(df[[var]], nrows,
-                                       dfs, var)
-    }
-    seen[matching] <- TRUE
-    if (all(seen))
-      break
-  }
-  list(setters = lapply(output, `[[`, "set"), getters = lapply(output, `[[`, "get"))
-}
-
-allocate_column <- function (example, nrows, dfs, var) {
-  a <- attributes(example)
-  type <- typeof(example)
-  class <- a$class
-  isList <- is.recursive(example)
-  a$names <- NULL
-  a$class <- NULL
-  if (is.data.frame(example)) {
-    stop("Data frame column '", var, "' not supported by rbind_fill")
-  }
-  if (is.array(example)) {
-    if (length(dim(example)) > 1) {
-      if ("dimnames" %in% names(a)) {
-        a$dimnames[1] <- list(NULL)
-        if (!is.null(names(a$dimnames)))
-          names(a$dimnames)[1] <- ""
-      }
-      df_has <- vapply(dfs, function(df) var %in% names(df),
-                       FALSE)
-      dims <- unique(lapply(dfs[df_has], function(df) dim(df[[var]])[-1]))
-      if (length(dims) > 1)
-        stop("Array variable ", var, " has inconsistent dims")
-      a$dim <- c(nrows, dim(example)[-1])
-      length <- prod(a$dim)
-    }
-    else {
-      a$dim <- NULL
-      a$dimnames <- NULL
-      length <- nrows
-    }
-  }
-  else {
-    length <- nrows
-  }
-  if (is.factor(example)) {
-    df_has <- vapply(dfs, function(df) var %in% names(df),
-                     FALSE)
-    isfactor <- vapply(dfs[df_has], function(df) is.factor(df[[var]]),
-                       FALSE)
-    if (all(isfactor)) {
-      levels <- unique(unlist(lapply(dfs[df_has], function(df) levels(df[[var]]))))
[…]
