+++ R/count_identifiers.R
@@ -20,7 +20,7 @@
 count_identifiers <- function(url = "http://oai.datacite.org/oai", prefix = 'oai_dc', ...) {
   check_url(url)
   args <- sc(list(verb = 'ListIdentifiers', metadataPrefix = prefix))
-  rbind_fill(lapply(url, ci, args = args, ...))
+  rbind.fill(lapply(url, ci, args = args, ...))
 }
 
 ci <- function(x, args, ...) {
