+++ R/oai-package.R
@@ -33,6 +33,7 @@
 #' @importFrom utils head
 #' @importFrom httr GET content stop_for_status
 #' @importFrom xml2 read_xml xml_children xml_text as_list xml_attrs xml_name xml_attr
+#' @importFrom plyr rbind.fill
 #' @docType package
 #' @title OAI-PMH Client
 #' @author Scott Chamberlain \email{myrmecocystus@@gmail.com}
