+++ DESCRIPTION
@@ -8,7 +8,7 @@ Description: A general purpose client to work with any 'OAI-PMH'
     Functions are provided to work with the 'OAI-PMH' verbs: 'GetRecord',
     'Identify', 'ListIdentifiers', 'ListMetadataFormats', 'ListRecords', and
     'ListSets'.
-Version: 0.1.1.9100
+Version: 0.1.1.9400
 License: MIT + file LICENSE
 Authors@R: c(
     person("Scott", "Chamberlain", role = c("aut", "cre"), email = "myrmecocystus@gmail.com"),
@@ -21,8 +21,9 @@ Imports:
     methods,
     stats,
     utils,
-    xml2,
-    httr (>= 1.0.0)
+    xml2 (>= 0.1.2),
+    httr (>= 1.0.0),
+    plyr (>= 1.8.3)
 Suggests:
     DBI,
     knitr,
