+++ NAMESPACE
@@ -28,6 +28,7 @@ importFrom(httr,GET)
 importFrom(httr,content)
 importFrom(httr,stop_for_status)
 importFrom(methods,is)
+importFrom(plyr,rbind.fill)
 importFrom(stats,setNames)
 importFrom(utils,head)
 importFrom(xml2,as_list)
