### 2015-10-21 ca9078e
change internal fxn to avoid conflict, #30
