+++ man/tsum.Rd
@@ -1,0 +1,43 @@
+% Generated by roxygen2 (4.1.1): do not edit by hand
+% Please edit documentation in R/tbldf.R
+\name{tsum}
+\alias{tsum}
+\alias{tsum.Date}
+\alias{tsum.POSIXt}
+\alias{tsum.array}
+\alias{tsum.character}
+\alias{tsum.default}
+\alias{tsum.factor}
+\alias{tsum.integer}
+\alias{tsum.logical}
+\alias{tsum.matrix}
+\alias{tsum.numeric}
+\title{Type summary}
+\usage{
+tsum(x)
+
+\method{tsum}{default}(x)
+
+\method{tsum}{character}(x)
+
+\method{tsum}{Date}(x)
+
+\method{tsum}{factor}(x)
+
+\method{tsum}{integer}(x)
+
+\method{tsum}{logical}(x)
+
+\method{tsum}{array}(x)
+
+\method{tsum}{matrix}(x)
+
+\method{tsum}{numeric}(x)
+
+\method{tsum}{POSIXt}(x)
+}
+\description{
+Type summary
+}
+\keyword{internal}
+
