+++ man/type_sum.Rd
@@ -1,43 +1,0 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/tbldf.R
-\name{type_sum}
-\alias{type_sum}
-\alias{type_sum.Date}
-\alias{type_sum.POSIXt}
-\alias{type_sum.array}
-\alias{type_sum.character}
-\alias{type_sum.default}
-\alias{type_sum.factor}
-\alias{type_sum.integer}
-\alias{type_sum.logical}
-\alias{type_sum.matrix}
-\alias{type_sum.numeric}
-\title{Type summary}
-\usage{
-type_sum(x)
-
-\method{type_sum}{default}(x)
-
-\method{type_sum}{character}(x)
-
-\method{type_sum}{Date}(x)
-
-\method{type_sum}{factor}(x)
-
-\method{type_sum}{integer}(x)
-
-\method{type_sum}{logical}(x)
-
-\method{type_sum}{array}(x)
-
-\method{type_sum}{matrix}(x)
-
-\method{type_sum}{numeric}(x)
-
-\method{type_sum}{POSIXt}(x)
-}
-\description{
-Type summary
-}
-\keyword{internal}
-
