+++ NAMESPACE
@@ -1,16 +1,16 @@
 # Generated by roxygen2 (4.1.1): do not edit by hand
 
 S3method(print,oai_df)
-S3method(type_sum,Date)
-S3method(type_sum,POSIXt)
-S3method(type_sum,array)
-S3method(type_sum,character)
-S3method(type_sum,default)
-S3method(type_sum,factor)
-S3method(type_sum,integer)
-S3method(type_sum,logical)
-S3method(type_sum,matrix)
-S3method(type_sum,numeric)
+S3method(tsum,Date)
+S3method(tsum,POSIXt)
+S3method(tsum,array)
+S3method(tsum,character)
+S3method(tsum,default)
+S3method(tsum,factor)
+S3method(tsum,integer)
+S3method(tsum,logical)
+S3method(tsum,matrix)
+S3method(tsum,numeric)
 export(count_identifiers)
 export(dump_raw_to_db)
 export(dump_raw_to_txt)
@@ -22,7 +22,7 @@ export(list_metadataformats)
 export(list_records)
 export(list_sets)
 export(load_providers)
-export(type_sum)
+export(tsum)
 export(update_providers)
 importFrom(httr,GET)
 importFrom(httr,content)
