### 2015-10-21 120e8c0
comment about where internal fxn comes from fix #31
