### 2015-10-21 a4575bd
change id to id_entify, fix #30, ubmped 99 ver, udpated readme
