+++ man/id_entify.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
 % Please edit documentation in R/id.R
-\name{id}
-\alias{id}
+\name{id_entify}
+\alias{id_entify}
 \title{Identify the OAI-PMH service for each data provider.}
 \usage{
-id(url, ...)
+id_entify(url, ...)
 }
 \arguments{
 \item{url}{OAI-PMH base url. Required.}
@@ -17,14 +17,14 @@ Identify the OAI-PMH service for each data provider.
 \examples{
 \dontrun{
 # datacite
-id("http://oai.datacite.org/oai")
+id_entify("http://oai.datacite.org/oai")
 
 # arxiv
-id("http://export.arxiv.org/oai2")
+id_entify("http://export.arxiv.org/oai2")
 
 # curl options
 library("httr")
-id("http://export.arxiv.org/oai2", config = verbose())
+id_entify("http://export.arxiv.org/oai2", config = verbose())
 }
 }
 
