+++ NAMESPACE
@@ -16,7 +16,7 @@ export(dump_raw_to_db)
 export(dump_raw_to_txt)
 export(dump_to_rds)
 export(get_records)
-export(id)
+export(id_entify)
 export(list_identifiers)
 export(list_metadataformats)
 export(list_records)
