### 2015-10-21 #19 @sckott
all changes made
