### 2015-10-21 #19 @sckott
closing - and moving into ropensci now
